Moved the ScheduleWizard from Common to ScheduleWizard folder
