+++ app/components/ScheduleWizard/ScheduleWizard.js
@@ -1,11 +1,11 @@
 import React, { Component } from 'react';
 import PropTypes from 'prop-types';
 import { observer, inject } from 'mobx-react';
-import TimeSettings from '../ScheduleWizard/TimeSettings';
-import InfoSettings from '../ScheduleWizard/InfoSettings';
-import BountySettings from '../ScheduleWizard/BountySettings';
-import ConfirmSettings from '../ScheduleWizard/ConfirmSettings';
-import PoweredByEAC from './PoweredByEAC';
+import TimeSettings from './TimeSettings';
+import InfoSettings from './InfoSettings';
+import BountySettings from './BountySettings';
+import ConfirmSettings from './ConfirmSettings';
+import PoweredByEAC from '../Common/PoweredByEAC';
 
 @inject('web3Service')
 @inject('scheduleStore')
